requirements.txt
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
dj-rest-auth==1.1.1
Django==3.1
django-allauth==0.42.0
django-phonenumber-field==5.0.0
django-restframework==0.0.1
djangorestframework==3.11.1
idna==2.10
ipdb==0.13.3
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
oauthlib==3.1.0
parso==0.7.1
pexpect==4.8.0
phonenumbers==8.12.8
pickleshare==0.7.5
prompt-toolkit==3.0.6
ptyprocess==0.6.0
Pygments==2.6.1
python3-openid==3.2.0
pytz==2020.1

